requirements.txt
django==4.2.11
djangorestframework==3.15.1
pillow==10.3.0